test(tau-coding-agent): split commands_and_packages gateway/deployment tests (#1383)
